ai-service/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
numpy
pandas
scikit-learn
tensorflow
torch
nltk
apscheduler
requests
mysql-connector-python
python-multipart
